Updates to add a main driver and single executable
