+++ pydfnworks/pydfnworks/general/paths.py
@@ -207,17 +207,12 @@ def define_paths(self):
          self.compile_dfn_exe(os.environ['dfnworks_PATH'] + 'DFNTrans/')
     self.valid('DFNTrans', os.environ['DFNTRANS_EXE'], "executable")
 
-    os.environ['CORRECT_UGE_EXE'] = os.environ[
-        'dfnworks_PATH'] + 'CPP_correct/correct_uge'
-    if not os.path.isfile(os.environ['CORRECT_UGE_EXE']):
-         self.compile_dfn_exe(os.environ['dfnworks_PATH'] + 'CPP_correct/')
-    self.valid('CORRECT_UGE_EXE', os.environ['CORRECT_UGE_EXE'], "executable")
-
-    os.environ['CORRECT_STOR_EXE'] = os.environ[
-        'dfnworks_PATH'] + 'CPP_correct/correct_stor'
-    if not os.path.isfile(os.environ['CORRECT_STOR_EXE']):
-         self.compile_dfn_exe(os.environ['dfnworks_PATH'] + 'CPP_correct/')
-    self.valid('CORRECT_STOR_EXE', os.environ['CORRECT_STOR_EXE'], "executable")
+    os.environ['CORRECT_VOLUME_EXE'] = os.environ[
+        'dfnworks_PATH'] + 'CPP_correct_volumes/correct_volume'
+    if not os.path.isfile(os.environ['CORRECT_VOLUME_EXE']):
+         self.compile_dfn_exe(os.environ['dfnworks_PATH'] +
+                        'CPP_correct_volumes/')
+    self.valid('CORRECT_VOLUME_EXE', os.environ['CORRECT_VOLUME_EXE'], "executable")
 
     os.environ['CONNECT_TEST_EXE'] = os.environ[
         'dfnworks_PATH'] + 'DFN_Mesh_Connectivity_Test/ConnectivityTest'
